requirements.txt
# Core dependencies
# Note: cwltool is optional but recommended for full CWL parsing
# Install with: pip install cwltool
# Or: pip install cwltool>=3.1.20231201152605

# YAML processing (ruamel.yaml is preferred, but pyyaml works as fallback)
ruamel.yaml>=0.17.32
pyyaml>=6.0.1

# HTTP and AWS
requests>=2.31.0
boto3>=1.34.0
botocore>=1.34.0

# CWL processing (optional - install if you need full CWL validation)
# cwl-utils>=0.25
# schema-salad>=8.4.20231201152605
cwltool>=3.1.20231201152605

# Nextflow generation
jinja2>=3.1.2
pyyaml>=6.0.1
nextflow>=22.04.0

# AWS HealthOmics
amazon-omics-tools

# Validation and testing
pytest>=7.4.3
pytest-cov>=4.1.0
jsonschema>=4.20.0

# Utilities
click>=8.1.7
tqdm>=4.66.1
colorama>=0.4.6
rich>=13.7.0

# Development
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1
pre-commit>=3.5.0

# Container handling
docker>=6.1.3
singularity>=0.1.0

# Logging and monitoring
structlog>=23.2.0
python-json-logger>=2.0.7
